+++ src/brewerypi/services/__init__.py
@@ -48,6 +48,13 @@ from brewerypi.services.enterprises import (
     list_enterprises,
     update_enterprise,
 )
+from brewerypi.services.event_frame_templates import (
+    create_event_frame_template,
+    delete_event_frame_template,
+    get_event_frame_template,
+    list_event_frame_templates,
+    update_event_frame_template,
+)
 from brewerypi.services.exceptions import (
     ConflictError,
     NotFoundError,
@@ -106,6 +113,7 @@ __all__ = [
     "create_element_attribute_template",
     "create_element_template",
     "create_enterprise",
+    "create_event_frame_template",
     "create_lookup",
     "create_lookup_value",
     "create_measurement_unit",
@@ -116,6 +124,7 @@ __all__ = [
     "delete_element_attribute_template",
     "delete_element_template",
     "delete_enterprise",
+    "delete_event_frame_template",
     "delete_lookup",
     "delete_lookup_value",
     "delete_measurement_unit",
@@ -128,6 +137,7 @@ __all__ = [
     "get_element_attribute_template",
     "get_element_template",
     "get_enterprise",
+    "get_event_frame_template",
     "get_lookup",
     "get_lookup_value",
     "get_measurement_unit",
@@ -140,6 +150,7 @@ __all__ = [
     "list_element_templates",
     "list_elements",
     "list_enterprises",
+    "list_event_frame_templates",
     "list_lookup_values",
     "list_lookups",
     "list_measurement_units",
@@ -151,6 +162,7 @@ __all__ = [
     "update_element_attribute_template",
     "update_element_template",
     "update_enterprise",
+    "update_event_frame_template",
     "update_lookup",
     "update_lookup_value",
     "update_measurement_unit",